Update and rename ReportingCheater.cs to Reporting.cs
