--- Exiled.Events/Patches/Events/Server/ReportingCheater.cs
@@ -1,71 +1,0 @@
-// -----------------------------------------------------------------------
-// <copyright file="ReportingCheater.cs" company="Exiled Team">
-// Copyright (c) Exiled Team. All rights reserved.
-// Licensed under the CC BY-SA 3.0 license.
-// </copyright>
-// -----------------------------------------------------------------------
-
-namespace Exiled.Events.Patches.Events.Server
-{
-#pragma warning disable SA1118
-    using System;
-    using System.Collections.Generic;
-    using System.Reflection.Emit;
-
-    using Exiled.Events.EventArgs;
-    using Exiled.Events.Handlers;
-
-    using HarmonyLib;
-
-    using NorthwoodLib.Pools;
-
-    using static HarmonyLib.AccessTools;
-
-    /// <summary>
-    /// Patches <see cref="CheaterReport.IssueReport(GameConsoleTransmission, string, string, string, string, string, string, ref string, ref byte[], string, int, string, string)"/>.
-    /// Adds the <see cref="Server.ReportingCheater"/> event.
-    /// </summary>
-    [HarmonyPatch(typeof(CheaterReport), nameof(CheaterReport.IssueReport))]
-    internal static class ReportingCheater
-    {
-        private static IEnumerable<CodeInstruction> Transpiler(IEnumerable<CodeInstruction> instructions, ILGenerator generator)
-        {
-            List<CodeInstruction> newInstructions = ListPool<CodeInstruction>.Shared.Rent(instructions);
-
-            var offset = 1;
-
-            var index = newInstructions.FindIndex(instruction => instruction.opcode == OpCodes.Leave) + offset;
-
-            var ev = generator.DeclareLocal(typeof(ReportingCheaterEventArgs));
-
-            var returnLabel = generator.DefineLabel();
-
-            newInstructions.InsertRange(index, new[]
-            {
-                new CodeInstruction(OpCodes.Ldarg_2),
-                new CodeInstruction(OpCodes.Call, Method(typeof(API.Features.Player), nameof(API.Features.Player.Get), new[] { typeof(string) })),
-                new CodeInstruction(OpCodes.Ldarg_3),
-                new CodeInstruction(OpCodes.Call, Method(typeof(API.Features.Player), nameof(API.Features.Player.Get), new[] { typeof(string) })),
-                new CodeInstruction(OpCodes.Ldsfld, Field(typeof(ServerConsole), nameof(ServerConsole.Port))),
-                new CodeInstruction(OpCodes.Ldarg_S, 4),
-                new CodeInstruction(OpCodes.Ldc_I4_1),
-                new CodeInstruction(OpCodes.Newobj, GetDeclaredConstructors(typeof(ReportingCheaterEventArgs))[0]),
-                new CodeInstruction(OpCodes.Dup),
-                new CodeInstruction(OpCodes.Dup),
-                new CodeInstruction(OpCodes.Stloc_S, ev.LocalIndex),
-                new CodeInstruction(OpCodes.Call, Method(typeof(Handlers.Server), nameof(Handlers.Server.OnReportingCheater))),
-                new CodeInstruction(OpCodes.Callvirt, PropertyGetter(typeof(ReportingCheaterEventArgs), nameof(ReportingCheaterEventArgs.IsAllowed))),
-                new CodeInstruction(OpCodes.Brfalse_S, returnLabel),
-                new CodeInstruction(OpCodes.Ldloc_S, ev.LocalIndex),
-                new CodeInstruction(OpCodes.Starg_S, 4),
-            });
-
-            newInstructions[newInstructions.Count - 1].WithLabels(returnLabel);
-
-            for (int z = 0; z < newInstructions.Count; z++)
-                yield return newInstructions[z];
-
-            ListPool<CodeInstruction>.Shared.Return(newInstructions);
-        }
-    }
-}
